Added infinite token functionality

+++ server/src/service/authorizedOnly.js
@@ -1,6 +1,6 @@
 const tokensCollection = require('../database/database.js')('tokens');
 
-// Доступ к ресурсам, имеющим этот middleware, могут получить толька авторизованные пользователи
+// Доступ к ресурсам, имеющим этот middleware, могут получить только авторизованные пользователи
 module.exports = async function authorizedOnly(req, res, next) {
     let tokenData = await tokensCollection.find({token: req.headers.authorization});
 
@@ -10,7 +10,7 @@ module.exports = async function authorizedOnly(req, res, next) {
         error.status = 401;
         next(error);
     } else {
-        if (tokenData.maxAge < Date.now()) {
+        if (tokenData.maxAge && tokenData.maxAge < Date.now()) {
             // Срок жизни токена истек
             await tokensCollection.deleteOne({token: tokenData.token});
             let error = new Error('Данный токен больше недействителен');
